+++ DX_Renderer/Core/Components/Resource/DescriptorHeap.hpp
@@ -38,7 +38,7 @@ namespace DXR
 		D3D12_CPU_DESCRIPTOR_HANDLE GetCPUHeapHandle() const;
 		D3D12_GPU_DESCRIPTOR_HANDLE GetGPUHeapHandle() const;
 	private:
-		DescriptorHeap(GraphicsDevice& device, UINT8 descriptorCount, DescriptorType type, D3D12_DESCRIPTOR_HEAP_FLAGS flags = D3D12_DESCRIPTOR_HEAP_FLAG_NONE);
+		DescriptorHeap(GraphicsDevice& device, UINT descriptorCount, DescriptorType type, D3D12_DESCRIPTOR_HEAP_FLAGS flags = D3D12_DESCRIPTOR_HEAP_FLAG_NONE);
 		inline void CreateDescriptorHeap(GraphicsDevice& device, D3D12_DESCRIPTOR_HEAP_FLAGS flags);
 	};
